in prog

+++ addons/godot-rollback-3d/rust/src/network/peer/peer_map.rs
@@ -35,10 +35,10 @@ pub fn apply_peer_map_to_network(network: &mut Network, peer_map: PeerMap) -> Re
 
     for (idx, peer_id) in peer_map.iter_shared().enumerate() {
         if peer_id == local_id {
-            network.local_peer.metadata = Some(PeerMetadata::new_with_idx(peer_id, idx as i64));
+            network.local_peer.metadata = Some(PeerMetadata::new_with_idx(peer_id, idx as u8));
         } else if let Some(peer) = network.get_remote_peer_mut(peer_id) {
             peer.metadata.id = peer_id;
-            peer.metadata.idx = Some(idx as i64);
+            peer.metadata.idx = Some(idx as u8);
         } else {
             log::error!("Local/remote peer with ID {} not found", peer_id);
         }
